TransmissionMechanisms

Documentation for TransmissionMechanisms.

TransmissionMechanisms.Q

Base.:!

Base.:&

Base.:|

Base.string

TransmissionMechanisms.apply_and!

TransmissionMechanisms.apply_not!

TransmissionMechanisms.calculate_Q_and_only

TransmissionMechanisms.check_contradiction

TransmissionMechanisms.collect_terms

TransmissionMechanisms.get_varnums_and_multiplier

TransmissionMechanisms.make_condition

TransmissionMechanisms.make_structural_B

TransmissionMechanisms.make_structural_Qbb

TransmissionMechanisms.remove_contradictions

TransmissionMechanisms.string_and

TransmissionMechanisms.to_structural_transmission_model

TransmissionMechanisms.to_transmission_irfs

TransmissionMechanisms.transmission

TransmissionMechanisms.transmission

TransmissionMechanisms.transmission

TransmissionMechanisms.Q — Type

Q(s::String)
Q(s::String, m::Number)
Q(s::Vector{String})
Q(s::Vector{String}, m::Vector{Number})
Q(i::Int)

Represents a transmission condition.

We denote with Q(b), where b is a Boolean statement, a transmission question.

Important: Each string in s: should be a Boolean statement involving variables xfollowed by a number, i.e.x1,x2`, etc. Additionally, each Boolean statement should contain only AND (&) and NOT (!) statements.

Important: Users should only use the Q(i::Int) constructor. All other constructors are for internal use only. Misuse of the other constructors easily leads to mistakes.

Fields

vars::Vector{String}: Contains the variables. These will be Boolean statements containing only AND and NOT.

multiplier::Vector{Number}: Multiplier in front of Q(b).

Arguments

s::Union{String, Vector{String}}: String representation of transmission condition.

m::Union{Number, Vector{Number}}: Multipliers for transmission conditions.

i::Int: Variable number.

Usage

# Defining all variables in one go
x = [Q("x$i") for i = 1:10]
q = (x[1] | x[2]) & !x[3]

# Alternatively variables can be defined separaterly
x1 = Q("x1")
x2 = Q("x2")
x3 = Q("x3")
q = (x1 | x2) & !x3

# The following are also valid
q = Q("x1 & !x3", 1)
q = Q(["x1", "x2", "x1 & x2"], [1, 1, -1])

# The following is NOT valid but does not yet throw an error or warning
q = Q("x1 | x2")  # DO NOT DO THIS!

Base.:! — Method

Base.!(q1::Q)

Return NOT the transmission condition if the condition involves more than one variables. If the condition only involves one variables, then "!x1" is returned where "1" is replaced by the respective variable number.

Note: The decision not to simplify terms of the form "!x1" was made because calculations usign the second calculation method in Wegner et al (2024) are faster than having to simplify "!x1" type of terms and using the first calculation method.

Base.:& — Method

Base.&(q1::Q, q2::Q)

Combine two transmission conditions using AND.

Base.:| — Method

Base.|(q1::Q, q2::Q)

Combine two transmission conditions using OR.

Base.string — Method

Base.string(q::Q)

Obtain a string representation of a transmission condition.

TransmissionMechanisms.apply_and! — Method

apply_and!(B::AbstractMatrix{T}, Qbb::AbstractMatrix{T}, from::Int, var::Int)

Manipulate B and Qbb so that var lies on all paths. This corresponds to zeroing out all edges going directly from the shock to any variables ordered after var and zeroing out any edges going from variables ordered before var to any variables ordered after var.

Arguments

B::AbstractMatrix{T}: Part of the structural transmission representation in Wegner et al (2024). See also make_structural_B.

Qbb::AbstractMatrix{T}: Part of the structural transmission representation in Wegner et al (2024). See also make_structural_Qbb.

from::Int: The shock number.

var::Int: The variable number that must lie on all paths.

Notes

This function is meant for internal use only.

TransmissionMechanisms.apply_not! — Method

apply_not!(B::AbstractMatrix{T}, Qbb::AbstractMatrix{T}, from::Int, var::Int)

Manipulate B and Qbb so that var lies on no paths. This corresponds to zeroing out the edge from the shock to var, and zeroing out all edges from variables ordered before var to var. The paper mentions also zeroing out edges leaving var, but this is not necessary.

Arguments

B::AbstractMatrix{T}: Part of the structural transmission representation in Wegner et al (2024). See also make_structural_B.

Qbb::AbstractMatrix{T}: Part of the structural transmission representation in Wegner et al (2024). See also make_structural_Qbb.

from::Int: The shock number.

var::Int: The variable number that must lie on all paths.

Notes

This function is meant for internal use only.

TransmissionMechanisms.calculate_Q_and_only — Method

calculate_Q_and_only(from::Int, 
    irfs::AbstractMatrix{T}, 
    irfs_ortho::AbstractMatrix{T}, 
    vars::AbstractVector{Int}, 
    multiplier::Number
) where {T}

Calculate the transmission effect of a transmission condition/query that involves only ANDs.

Arguments

from::Int: ID of shock.

irfs::AbstractMatrix{T}: IRFs in transmission form. See to_transmission_irfs.

irfs_ortho::AbstractMatrix{T}: Orthogonalised IRFs in transmission form. See to_transmission_irfs.

multiplier::Number: Multiplier.

Returns

Returns a Vector{T} with entry i corresponding to the transmission effect on variable $x_i$. If $x_k$ is the variable in the transmission condition with the highest subscript, then all entries in the returned vector with index less thatn k are NaN since interpretation of those results is nonsensical.

Notes

Only for internal use.

TransmissionMechanisms.check_contradiction — Method

check_contradiction(var_and::Vector{Int}, var_not::Vector{Int})
check_contradiction(var_and::Vector{Vector{Int}}, var_not::Vector{Vector{Int}})

Check whether there is a contradiction of the form x1 & !x1.

Arguments

var_and::Union{Vector{Int}, Vector{Vector{Int}}}: AND variable numbers obtained from get_varnums_and_multiplier.

var_not::Union{Vector{Int}, Vector{Vector{Int}}}: NOT variable numbers obtained from get_varnums_and_multiplier

Returns

Bool indicating whether there are any contradictions.

Vector{Bool} indicating which elements yielded a contradiction.

Notes

This is used in remove_contradictions to remove contradicting terms. This speeds up simplification of terms, since the total number of terms can often be reduced.

TransmissionMechanisms.collect_terms — Method

collect_terms(q::Q)

Collect all terms Q(b) for which the Boolean statement b is the same and sums their multiplier. The result is a transmission condition for which each term only appears ones, but with multipliers possibly different from plus-minus one.

Arguments

q::Q: A transmission condition. See also Q

Returns

Another transmission condition of type Q.

Example

q = Q(["x1", "x1"], [1, 1])  
collect_terms(q)
# output: Q("x1", 2)
q = Q(["x1", "", "x1"], [1, 1, -1])  
collect_terms(q)
# output: Q("", 1)

TransmissionMechanisms.get_varnums_and_multiplier — Method

get_varnums_and_multiplier(q::Q)

Obtain the AND and NOT expressions of a transmission condition. Also return the multiplier for each term.

A valid transmission condition is a set of terms involving only AND and NOT expressions. For each term, the AND and NOT expressions are collected and a vector of the respective variable numbers is returned.

Arguments

q::Q: A transmission condition. See also Q.

Returns

and_nums::Vector{Vector{Int}}: Contains for each term a vector of variable numbers that are included via AND in the term.

and_not_nums::Vector{Vector{Int}}: Contains for each term a vector of variable numbers that are included via NOT in the term.

multiplier::Vector{Number}: Contains for each term the multiplier.

Examples

q = Q(["x1", "!x2", "x1 & !x2"], [1, 2, 3])
and_nums, and_not_nums, multiplier = get_varnums_and_multiplier(q)
# output: 
# and_nums = [[1], [], [1]]
# and_not_nums = [[], [2], [2]]
# multiplier = [1, 2, 3]

TransmissionMechanisms.make_condition — Method

make_condition(s::String)

Given a s::String, form a transmission condition of the form Q(b) where b is a Boolean statement.

Transmission mechanisms are described using Boolean statements involving the variables in the system. Each variable starts with x followed by a number. For example, given a three variable VAR(1), y{1,t} -> x1, y{2, t} -> x2, y{3, t} -> x3, y{1, t+1} -> x4, y{2, t+1} -> x5, ... Boolean statements then involve expressions in the x variables and define which paths can be taken. Each path involved in the transmission mechanism must satisfy the Boolean statement.

Arguments

s::String: A Boolean statement given as a string. Variables must start with x for them to be valid variables.

Returns

Returns a transmission condition. See also Q.

Examples

s = "x2 & !x3"
cond = make_condition(s)

Notes

Ensure that variables in the input string are correctly formatted as x followed by a number.

The resulting transmission condition can be used in transmission to calculate the transmission effect.

TransmissionMechanisms.make_structural_B — Method

make_structural_B(A0_ortho::AbstractMatrix, Aplus_ortho::AbstractMatrix, p::Int, max_horizon::Int)

Create the B matrix in the structural representation in WEGNER.

Given a SVAR(p) in the form of

\[y_t'A_0 = y_{t-1}'A_1 + \dots + y_{t-p}'A_p + \varepsilon_t' = [y_{t-1}', \dots, y_{t-p}']A^+ + \varepsilon_t',\]

and a pre-specified maximum horizon of $h$, the SVAR(p) over the next h horizons can be written as

\[x = Bx + \mathbb{Q}\epsilon,\]

with the structure of $B$ and $\mathbb{Q}$ given in the paper.

Arguments

A0_ortho::AbstractMatrix: The contemporaneous matrix of the SVAR(p) obtained using a Cholesky decomposition. Note that this would be an upper-triangular matrix in the representation chosen here.

Aplus_ortho::AbstractMatrix: The structural lag matrix of the SVAR(p) obtained using a Cholesky decomposition.

p::Int: The order of the SVAR(p).

max_horizon::Int: The maximum horizon to consider for the transmission model.

Returns

Returns $B$ in the above structural representation of the SVAR(p) over the next $h$ horizons.

Notes

See also make_structural_Qbb and to_structural_transmission_model.

TransmissionMechanisms.make_structural_Qbb — Method

make_structural_B(A0_ortho::AbstractMatrix, Aplus_ortho::AbstractMatrix, p::Int, max_horizon::Int)

Create the B matrix in the structural representation in WEGNER.

Given a SVAR(p) in the form of

\[y_t'A_0 = y_{t-1}'A_1 + \dots + y_{t-p}'A_p + \varepsilon_t' = [y_{t-1}', \dots, y_{t-p}']A^+ + \varepsilon_t',\]

and a pre-specified maximum horizon of $h$, the SVAR(p) over the next h horizons can be written as

\[x = Bx + \mathbb{Q}\epsilon,\]

with the structure of $B$ and $\mathbb{Q}$ given in the paper.

Arguments

irf0::AbstractMatrix: The identified impact responses. If the ith shock has not been identified, then the ith column consists of NaN. Matrix should be of dimension k times k with k being the number of variables in the SVAR(p).

A0_ortho::AbstractMatrix: The contemporaneous matrix of the SVAR(p) obtained using a Cholesky decomposition. Note that this would be an upper-triangular matrix in the representation chosen here.

max_horizon::Int: The maximum horizon to consider for the transmission model.

Returns

Returns $\mathbb{Q}$ in the above structural representation of the SVAR(p) over the next $h$ horizons.

Notes

See also make_structural_B and to_structural_transmission_model.

TransmissionMechanisms.remove_contradictions — Method

remove_contradictions(q::Q)

Remove contradicting terms.

A terms is deemed contradicting if it includes some "xi & !xi". This would result in the entire Boolean statement to be false, and thus in the effect of this terms to be zero.

Arguments

q::Q: A transmission condition. See also Q and make_condition.

Returns

If TransmissionMechanisms.REMOVE_CONTRADICTIONS == false, then q will simply be returned again.

If TransmissionMechanisms.REMOVE_CONTRADICTIONS == false, then

If all terms are contradicting, the Q("", 0) will be retuned, which has a transmission effect of zero.

If some terms are non-contradicting, then a transmission condition consisting of only the non-contradicting terms will be returned.

Examples

TransmissionMechanisms.REMOVE_CONTRADICTIONS = true
q = TransmissionMechanisms.Q("x1", 1)
remove_contradictions(q)  # will return q again since no contradictions exist

q = TransmissionMechanisms.Q("x1 & !x1", 1)
remove_contradictions(q)  # Will return Q("", 0)

q = TransmissionMechanisms.Q(["x1 & !x1", "x1 & x2"], [1, 1])
remove_contradictions(q)  # Will return Q("x1 & x2", 1)

TransmissionMechanisms.string_and — Method

string_and(s1::String, s2::String)

Combine two strings using "&".

TransmissionMechanisms.to_structural_transmission_model — Method

make_structural_B(A0_ortho::AbstractMatrix, Aplus_ortho::AbstractMatrix, p::Int, max_horizon::Int)

Create the B matrix in the structural representation in WEGNER.

Given a SVAR(p) in the form of

\[y_t'A_0 = y_{t-1}'A_1 + \dots + y_{t-p}'A_p + \varepsilon_t' = [y_{t-1}', \dots, y_{t-p}']A^+ + \varepsilon_t',\]

and a pre-specified maximum horizon of $h$, the SVAR(p) over the next h horizons can be written as

\[x = Bx + \mathbb{Q}\epsilon,\]

with the structure of $B$ and $\mathbb{Q}$ given in the paper.

Arguments

irf0::AbstractMatrix: The identified impact responses. If the ith shock has not been identified, then the ith column consists of NaN. Matrix should be of dimension k times k with k being the number of variables in the SVAR(p).

A0_ortho::AbstractMatrix: The contemporaneous matrix of the SVAR(p) obtained using a Cholesky decomposition. Note that this would be an upper-triangular matrix in the representation chosen here.

Aplus_ortho::AbstractMatrix: The structural lag matrix of the SVAR(p) obtained using a Cholesky decomposition.

p::Int: The order of the SVAR(p).

max_horizon::Int: The maximum horizon to consider for the transmission model.

Returns

Returns $B$ and $\mathbb{Q}$ in the above structural representation of the SVAR(p) over the next $h$ horizons in this order.

TransmissionMechanisms.to_transmission_irfs — Method

to_transmission_irfs(irfs::AbstractArray{T, 3})

Transform a standard three dimensional IRF array into a IRF matrix. The

Arguments

irfs::AbstractArray{T, 3}: IRF Array of dimension nvariabels × nshocks × n_horizons with the first horizons corresponding to horizon 0.

Returns

Matrix{T} of dimension (nvariables * nhorizons) × (nvariables * nhorizons). This is the same as what would be obtained via $(I-B)^{-1}\mathbb{Q}$ using the notation of Wegner et al (2024).

TransmissionMechanisms.transmission — Method

transmission(from::Int, 
    B::AbstractMatrix{T},
    Qbb::AbstractMatrix{T}, 
    q::Q, 
    ::Type{Val{:BQbb}}
) where {T}

Given a transmission condition q, calculate the transmission effect using the :BQbb method.

Arguments

from::Int: Shock number.

B::AbstractMatrix{T}: Part of the structural transmission representation in Wegner et al (2024). See also make_structural_B.

Qbb::AbstractMatrix{T}: Part of the structural transmission representation in Wegner et al (2024). See also make_structural_Qbb.

q::Q: A transmission condition. See also Q.

Returns

Returns a Vector{T} with entry i corresponding to the transmission effect on variable $x_i$. If $x_k$ is the variable in the transmission condition with the highest subscript, then all entries in the returned vector with index less thatn k are NaN since interpretation of those results is nonsensical.

Examples

k = 6
h = 3
s = "(x1 | x2) & !x3"
cond = make_condition(s)

B = randn(k*(h+1), k*(h+1))
Qbb = randn(k*(h+1), k*(h+1))

effect = transmission(1, B, Qbb, cond)

TransmissionMechanisms.transmission — Method

transmission(from::Int, 
    irfs::AbstractMatrix{T}, 
    irfs_ortho::AbstractMatrix{T}, 
    q::Q, 
    ::Type{Val{:irfs}}
) where {T}

Given a transmission condition q, calculate the transmission effect using the :irfs method.

Arguments

from::Int: Shock number.

irfs::AbstractMatrix{T}: Impulse responses. These should be in the form of the structural transmission model. See also to_transmission_irfs.

irfs_ortho::AbstractMatrix{T}: Orthogonalised IRFs. These should be in the form of the structural transmission model. See also to_transmission_irfs.

q::Q: A transmission condition. See also Q.

Returns

Returns a Vector{T} with entry i corresponding to the transmission effect on variable $x_i$. If $x_k$ is the variable in the transmission condition with the highest subscript, then all entries in the returned vector with index less thatn k are NaN since interpretation of those results is nonsensical.

Examples

k = 6
h = 3
s = "(x1 | x2) & !x3"
cond = make_condition(s)

irfs = randn(k, k, h+1)
irfs_ortho = randn(k, k, h+1)

irfs = to_transmission_irfs(irfs)
irfs_ortho = to_transmission_irfs(irfs_ortho)

effect = transmission(1, irfs, irfs_ortho, cond; method = :irfs)

TransmissionMechanisms.transmission — Method

transmission(from::Int, arr1::AbstractMatrix{T}, arr2::AbstractMatrix{T}, q::Q; method = :BQbb) where {T}

Given a transmission condition q, calculate the transmission effect using the either the :BQbb method (the default), or the :irfs method.

Arguments

from::Int: Shock number.

arr1::AbstractMatrix{T}. In case of :BQbb this must be B, in case of :irfs this must be irfs. See the documentation for the specific methods for transmission(..., ::Type{Val{:BQbb}}) and transmission(...,::Type{Val{:irfs}}).

arr2::AbstractMatrix{T}: In case of :BQbb this must be Qbb, in case of :irfs this must be irfs_ortho. See the documentation for the specific methods for transmission(..., ::Type{Val{:BQbb}}) and transmission(...,::Type{Val{:irfs}}).

q::Q: A transmission condition. See also Q.

Keyword Arguments

method::Symbol: Either :BQbb in which case the transmission effect will be calculated using the second method in Wegner et al (2024), or :irfs in which case the transmission effect is calculated using the first method in Wegner et al (2024).

Returns

Returns a Vector{T} with entry i corresponding to the transmission effect on variable $x_i$. If $x_k$ is the variable in the transmission condition with the highest subscript, then all entries in the returned vector with index less thatn k are NaN since interpretation of those results is nonsensical.

Examples

k = 6
h = 3
s = "(x1 | x2) & !x3"
cond = make_condition(s)

B = randn(k*(h+1), k*(h+1))
Qbb = randn(k*(h+1), k*(h+1))

effect = transmission(1, B, Qbb, cond; method = :BQbb)
effect = transmission(1, B, Qbb, cond)  # same as above; default is :BQbb

irfs = randn(k, k, h+1)
irfs_ortho = randn(k, k, h+1)

irfs = to_transmission_irfs(irfs)
irfs_ortho = to_transmission_irfs(irfs_ortho)

effect = transmission(1, irfs, irfs_ortho, cond; method = :irfs)