TransmissionMechanisms

Documentation for TransmissionMechanisms.

TransmissionMechanisms.collect_and_terms

TransmissionMechanisms.contains_nots

TransmissionMechanisms.create_transmission_function

TransmissionMechanisms.get_terms

TransmissionMechanisms.helper_Q

TransmissionMechanisms.helper_sym_to_num

TransmissionMechanisms.is_not_valid_variable_name

TransmissionMechanisms.make_condition

TransmissionMechanisms.make_structural_B

TransmissionMechanisms.make_structural_Qbb

TransmissionMechanisms.to_structural_transmission_model

TransmissionMechanisms.collect_and_terms — Method

collect_and_terms(expr::Union{SymbolicUtils.BasicSymbolic{Bool}, Vector{SymbolicUtils.BasicSymbolic{Bool}}}; rev=true)

Collect unique AND terms from a boolean expression or a vector of boolean expressions.

This function extracts unique AND terms from the input boolean expression(s) and returns them as a vector of SymbolicUtils.BasicSymbolic{Bool}. The input expr can be either a single boolean expression or a vector of boolean expressions.

Arguments

expr: A single boolean expression or a vector of boolean expressions from which AND terms will be collected.

rev::Bool=true: (Optional) If true, the resulting AND terms are sorted in descending order based on their string representation. If false, the sorting is in ascending order.

Returns

Returns a vector of SymbolicUtils.BasicSymbolic{Bool} containing unique AND terms extracted from the input expression(s).

Examples

julia s = "x2 & x3 & !x2 & x5" cond = make_condition(s) and_terms = TransmissionMechanisms.collect_and_terms(cond)`

TransmissionMechanisms.contains_nots — Method

contains_nots(term::SymbolicUtils.BasicSymbolic{Bool})

Check whether a term still contains a NOT statement.

To be able to calculate transmission effects using IRFs, Boolean conditions need to be simplified to a state in which they only involve ANDs in all terms. Thus, this function checks whether something went wrong in the simplification process.

Arguments

term::SymbolicUtils.BasicSymbolic{Bool}: A term as returned by helper_Q or get_terms.

Returns

Returns true if the term includes any NOT and false otherwise.

TransmissionMechanisms.create_transmission_function — Method

create_transmission_function(from::Int, [to::Int,] condition::SymbolicUtils.BasicSymbolic{Bool})

Create a transmission function based on a boolean condition.

This function generates a function that calculates the transmission effect from the fromth variables to the optional toth variable. If no to is given, the effect is calculated for all destination nodes. The transmission paths must satisfy the Boolean statement given in condition which can be created using make_condition.

Arguments

from::Int: Index of the variable from which the transmission starts.

[to::Int]: (Optional) Index of the variable to which the transmission ends. If omitted, the returned function calcuates the effect for all destination nodes.

condition::SymbolicUtils.BasicSymbolic{Bool}: Boolean condition specifying the transmission mechanism. The condition is represented as a boolean expression using variables starting with x and can be created using make_condition.

Returns

Returns a transmission function that can be applied to impulse response functions (irfs) and orthogonalized impulse response functions (irfs_ortho) to compute the transmission effects. The first argument in the returned function is irfs and the second arguement is irfs_ortho. Both should be matrices corresponding the the structural and orthogonalised IRFs respectively. Both should be square and of dimension k*(h+1) with h being the maximum horizon and k being the number of variables. The structural version can, for example, be obtained via irfs = inv(I - B) * Qbb, with B and Qbb obtained from to_structural_transmission_model.

Examples

julia s = "x2 & !x3" cond = make_condition(s) tf = create_transmission_function(1, cond) irfs = randn(3, 3) irfs_ortho = randn(3, 3) tf(irfs, irfs_ortho)`

TransmissionMechanisms.get_terms — Method

get_terms(condition::SymbolicUtils.BasicSymbolic{Bool})

Simplify a Boolean condition into a sum of only conjunctions.

Given a valid Boolean condition in the context of transmission mechanisms, the condition is recursively simplified until the result only consists of additive terms with each remaining condition being purely a conjunction. For example, The condition x2 & !x3 corresponds to the transmission statement Q(x2 & !x3) which is simplified to Q(x2) - Q(x2 & x3). The function will then return [Q(x2), -Q(x2 & x3)]. Each term can then be easily evaluated using IRFs.

Arguments

condition::SymbolicUtils.BasicSymbolic{Bool}: A symbolic boolean condition. Can be created using make_condition. Valid variable names are x followed by a number, or T respresenting true. All transmission paths must satisfy this boolean statement.

Returns

Returns a vector of terms, each of which consists of a conjunction.

Examples

s = "x2 & !x3"
cond = make_condition(s)
term = get_terms(cond)  # [Q(x2), -Q(x2 & x3)]

TransmissionMechanisms.helper_Q — Method

helper_Q(condition::SymbolicUtils.BasicSymbolic{Bool})

Simplifies the valid Boolean condition in the context of transmission mechanisms and returns all necessary information to evaluate the transmission effect using information in Impulse Response Functions (IRFs).

Arguments

condition::SymbolicUtils.BasicSymbolic{Bool}: A valid symbolic Boolean condition. Can be created using make_condition.

Returns

A vector containing all the additive terms in the simplified representation. Each term only involves conjunctions and can thus be calculated using information in IRFs.

A vector of multipliers applied to each term.

A vector of variables involved in the transmission query.

A vector of vectors. Each inner vector corresponds to a term and provides the variable numbers involved in the term. These, together with the multipliers, can be used to calculate the transmission effect.

TransmissionMechanisms.helper_sym_to_num — Method

helper_sym_to_num(sym_vec::Union{SymbolicUtils.BasicSymbolic{Bool}, Vector{SymbolicUtils.BasicSymbolic{Bool}}})

Convert symbolic variables to numerical indices.

This function takes a single boolean symbolic variable or an array of boolean symbolic variables sym_vec and converts them into numerical indices. Each index corresponds to the variable index in the structural transmission model and thus to an index in the IRF matrix. Valid variable names should start with 'x' followed by a number or be the symbol 'T'. If invalid variable names are encountered, an error is raised.

Arguments

sym_vec::Union{Any, Vector{Any}}: A single boolean symbolic variable or an array of boolean symbolic variables representing variables to be converted to numerical indices.

Returns

Returns a sorted vector of numerical indices corresponding to the input symbols. If a symbol is 'T', it is represented as nothing in the output vector.

Examples

```julia @syms x1::Bool x2::Bool x3::Bool T::Bool y::Bool TransmissionMechanisms.helpersymtonum(x2) # Returns [2] TransmissionMechanisms.helpersymtonum(T) # Returns [nothing] TransmissionMechanisms.helpersymtonum([x1, x3]) # Returns [1, 3] TransmissionMechanisms.helpersymtonum(y) # Error: Variable names should start with 'x' followed by a number.

TransmissionMechanisms.is_not_valid_variable_name — Method

is_not_valid_variable_name(sym::SymbolicUtils.BasicSymbolic{Bool})

Check if a given symbol is a valid variable name in the context of transmission mechanisms.

This function validates whether the provided symbol sym is a valid variable name for use in transmission mechanisms and in create_transmission_function. In the context of transmission mechanisms, valid variable names are of the form 'x' followed by one or more digits, or the symbol 'T' representing true.

Arguments

sym::SymbolicUtils.BasicSymbolic{Bool}: Symbol to be checked for validity as a variable name in transmission mechanisms.

Returns

Returns true if the input symbol is not a valid variable name, and false otherwise.

Examples

```julia @syms x1::Bool T::Bool y::Bool TransmissionMechanisms.isnotvalidvariablename(x1) # Returns false (valid variable name) TransmissionMechanisms.isnotvalidvariablename(y) # Returns true (invalid variable name) TransmissionMechanisms.isnotvalidvariablename(T) # Returns false (valid variable name)

TransmissionMechanisms.make_condition — Method

make_condition(string::String)

Given a string, make a SymbolicUtils.jl Boolean expression.

Transmission mechanisms are described using Boolean statements involving the variables in the system. Each variable starts with x followed by a number. For example, given a three variable VAR(1), y{1,t} -> x1, y{2, t} -> x2, y{3, t} -> x3, y{1, t+1} -> x4, y{2, t+1} -> x5, ... Boolean statements then involve expressions in the x variables and define which paths can be taken. Each path involved in the transmission mechanism must satisfy the Boolean statement.

Arguments

string::String: A Boolean statement given as a string. Variables must start with x for them to be valid variables.

Returns

Returns a SymbolicUtils.jl Boolean expression that can be used in create_transmission_function.

Examples

s = "x2 & !x3"
cond = make_condition(s)
tf = create_transmission_function(1, cond)
irfs = randn(3, 3)
irfs_ortho = randn(3, 3)
tf(irfs, irfs_ortho)

Notes

Ensure that variables in the input string are correctly formatted as x followed by a number.

The resulting Boolean expression can be used in create_transmission_function to create a function that calculates the transmission effect.

TransmissionMechanisms.make_structural_B — Method

make_structural_B(A0_ortho::AbstractMatrix, Aplus_ortho::AbstractMatrix, p::Int, max_horizon::Int)

Create the B matrix in the structural representation in Wegner et al (2024).

Given a SVAR(p) in the form of

\[y_t'A_0 = y_{t-1}'A_1 + \dots + y_{t-p}'A_p + \varepsilon_t' = [y_{t-1}', \dots, y_{t-p}']A^+ + \varepsilon_t',\]

and a pre-specified maximum horizon of $h$, the SVAR(p) over the next h horizons can be written as

\[x = Bx + \mathbb{Q}\epsilon,\]

with the structure of $B$ and $\mathbb{Q}$ given in the paper.

Arguments

A0_ortho::AbstractMatrix: The contemporaneous matrix of the SVAR(p) obtained using a Cholesky decomposition. Note that this would be an upper-triangular matrix in the representation chosen here.

Aplus_ortho::AbstractMatrix: The structural lag matrix of the SVAR(p) obtained using a Cholesky decomposition.

p::Int: The order of the SVAR(p).

max_horizon::Int: The maximum horizon to consider for the transmission model.

Returns

Returns $B$ in the above structural representation of the SVAR(p) over the next $h$ horizons.

Notes

See also make_structural_Qbb and to_structural_transmission_model.

TransmissionMechanisms.make_structural_Qbb — Method

make_structural_B(A0_ortho::AbstractMatrix, Aplus_ortho::AbstractMatrix, p::Int, max_horizon::Int)

Create the B matrix in the structural representation in Wegner et al (2024).

Given a SVAR(p) in the form of

\[y_t'A_0 = y_{t-1}'A_1 + \dots + y_{t-p}'A_p + \varepsilon_t' = [y_{t-1}', \dots, y_{t-p}']A^+ + \varepsilon_t',\]

and a pre-specified maximum horizon of $h$, the SVAR(p) over the next h horizons can be written as

\[x = Bx + \mathbb{Q}\epsilon,\]

with the structure of $B$ and $\mathbb{Q}$ given in the paper.

Arguments

irf0::AbstractMatrix: The identified impact responses. If the ith shock has not been identified, then the ith column consists of NaN. Matrix should be of dimension k times k with k being the number of variables in the SVAR(p).

A0_ortho::AbstractMatrix: The contemporaneous matrix of the SVAR(p) obtained using a Cholesky decomposition. Note that this would be an upper-triangular matrix in the representation chosen here.

max_horizon::Int: The maximum horizon to consider for the transmission model.

Returns

Returns $\mathbb{Q}$ in the above structural representation of the SVAR(p) over the next $h$ horizons.

Notes

See also make_structural_B and to_structural_transmission_model.

TransmissionMechanisms.to_structural_transmission_model — Method

make_structural_B(A0_ortho::AbstractMatrix, Aplus_ortho::AbstractMatrix, p::Int, max_horizon::Int)

Create the B matrix in the structural representation in Wegner et al (2024).

Given a SVAR(p) in the form of

\[y_t'A_0 = y_{t-1}'A_1 + \dots + y_{t-p}'A_p + \varepsilon_t' = [y_{t-1}', \dots, y_{t-p}']A^+ + \varepsilon_t',\]

and a pre-specified maximum horizon of $h$, the SVAR(p) over the next h horizons can be written as

\[x = Bx + \mathbb{Q}\epsilon,\]

with the structure of $B$ and $\mathbb{Q}$ given in the paper.

Arguments

irf0::AbstractMatrix: The identified impact responses. If the ith shock has not been identified, then the ith column consists of NaN. Matrix should be of dimension k times k with k being the number of variables in the SVAR(p).

A0_ortho::AbstractMatrix: The contemporaneous matrix of the SVAR(p) obtained using a Cholesky decomposition. Note that this would be an upper-triangular matrix in the representation chosen here.

Aplus_ortho::AbstractMatrix: The structural lag matrix of the SVAR(p) obtained using a Cholesky decomposition.

p::Int: The order of the SVAR(p).

max_horizon::Int: The maximum horizon to consider for the transmission model.

Returns

Returns $B$ and $\mathbb{Q}$ in the above structural representation of the SVAR(p) over the next $h$ horizons in this order.